requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
cryptography==3.1.1
decorator==4.4.2
feedparser==6.0.1
pycparser==2.20
python-telegram-bot==13.0
pytz==2020.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
tzlocal==2.1
